+++ glui/include/GL/glui.h
@@ -1411,251 +1411,6 @@ namespace glui {
   class GLUI_Rotation;
   class GLUI_Translation;
 
-  /**
-     The main user-visible interface object to GLUI_Context *.
- 
-  */
-  class GLUI_API GLUI_Context : public GLUI_Main 
-  {
-  public:
-    /** DEPRECATED interface for creating new GLUI_Context *objects */
-    int   add_control( GLUI_Control *control ) { return main_panel->add_control(control); }
-
-    void  add_column( int draw_bar = true );
-    void  add_column_to_panel( GLUI_Panel *panel, int draw_bar = true );
-
-    void  add_separator( void );
-    void  add_separator_to_panel( GLUI_Panel *panel );
-
-    GLUI_RadioGroup 
-      *add_radiogroup( int *live_var=NULL,
-                       int user_id=-1,CallBack callback=CallBack());
-
-    GLUI_RadioGroup 
-      *add_radiogroup_to_panel(  GLUI_Panel *panel,
-                                 int *live_var=NULL,
-                                 int user_id=-1, CallBack callback=CallBack() );
-    GLUI_RadioButton
-      *add_radiobutton_to_group(  GLUI_RadioGroup *group,
-                                  const char *name );
-
-    GLUI_Listbox *add_listbox( const char *name, int *live_var=NULL,
-                               int id=-1, CallBack callback=CallBack()	);
-    GLUI_Listbox *add_listbox_to_panel( GLUI_Panel *panel,
-                                        const char *name, int *live_var=NULL,
-                                        int id=-1, CallBack callback=CallBack());
-
-    GLUI_Rotation *add_rotation( const char *name, float *live_var=NULL,
-                                 int id=-1, CallBack callback=CallBack()	);
-    GLUI_Rotation *add_rotation_to_panel( GLUI_Panel *panel,
-                                          const char *name, float *live_var=NULL,
-                                          int id=-1, CallBack callback=CallBack());
-  
-    GLUI_Translation *add_translation( const char *name,
-                                       int trans_type, float *live_var=NULL,
-                                       int id=-1, CallBack callback=CallBack()	);
-    GLUI_Translation *add_translation_to_panel( 
-                                               GLUI_Panel *panel, const char *name, 
-                                               int trans_type, float *live_var=NULL,
-                                               int id=-1, CallBack callback=CallBack());
-  
-    GLUI_Checkbox  *add_checkbox( const char *name, 
-                                  int *live_var=NULL,
-                                  int id=-1, CallBack callback=CallBack());
-    GLUI_Checkbox  *add_checkbox_to_panel( GLUI_Panel *panel, const char *name, 
-                                           int *live_var=NULL, int id=-1, 
-                                           CallBack callback=CallBack());
-
-    GLUI_Button  *add_button( const char *name, int id=-1, 
-                              CallBack callback=CallBack());
-    GLUI_Button  *add_button_to_panel( GLUI_Panel *panel, const char *name, 
-                                       int id=-1, CallBack callback=CallBack() );
-
-    GLUI_StaticText  *add_statictext( const char *name );
-    GLUI_StaticText  *add_statictext_to_panel( GLUI_Panel *panel, const char *name );
-
-    GLUI_EditText  *add_edittext( const char *name, 
-                                  int data_type=GLUI_EDITTEXT_TEXT,
-                                  void*live_var=NULL,
-                                  int id=-1, CallBack callback=CallBack()	);
-    GLUI_EditText  *add_edittext_to_panel( GLUI_Panel *panel, 
-                                           const char *name,
-                                           int data_type=GLUI_EDITTEXT_TEXT,
-                                           void *live_var=NULL, int id=-1, 
-                                           CallBack callback=CallBack() );
-    GLUI_EditText  *add_edittext( const char *name, GLUI_String& live_var, 
-                                  int id=-1, CallBack callback=CallBack()	);
-    GLUI_EditText  *add_edittext_to_panel( GLUI_Panel *panel, const char *name, 
-                                           GLUI_String& live_var, int id=-1,
-                                           CallBack callback=CallBack() );
-
-    GLUI_Spinner  *add_spinner( const char *name, 
-                                int data_type=GLUI_SPINNER_INT,
-                                void *live_var=NULL,
-                                int id=-1, CallBack callback=CallBack() );
-    GLUI_Spinner  *add_spinner_to_panel( GLUI_Panel *panel, 
-                                         const char *name,
-                                         int data_type=GLUI_SPINNER_INT,
-                                         void *live_var=NULL,
-                                         int id=-1,
-                                         CallBack callback=CallBack() );
-
-    GLUI_Panel     *add_panel( const char *name, int type=GLUI_PANEL_EMBOSSED );
-    GLUI_Panel     *add_panel_to_panel( GLUI_Panel *panel, const char *name, 
-                                        int type=GLUI_PANEL_EMBOSSED );
-
-
-    GLUI_Rollout   *add_rollout( const char *name, int open=true,
-                                 int type=GLUI_PANEL_EMBOSSED);
-    GLUI_Rollout   *add_rollout_to_panel( GLUI_Panel *panel, const char *name, 
-                                          int open=true,
-                                          int type=GLUI_PANEL_EMBOSSED);
-
-
-    /** Set the window where our widgets should be displayed. */
-    void            set_main_gfx_window( int window_id );
-    int             get_glut_window_id( void ) { return glut_window_id; }
-
-    void            enable( void ) { main_panel->enable(); }
-    void            disable( void );
-
-    void            sync_live( void );
-
-    void            close( void );
-
-    void            show( void );
-    void            hide( void );
-
-    /***** Constructors and desctructors *****/
-
-    int init( const char *name, long flags, int x, int y, int parent_window );
-  protected:
-    virtual int add_control( GLUI_Node *parent, GLUI_Control *control ) {
-      return GLUI_Main::add_control( parent, control );
-    }
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               EditText class                             */
-  /*                                                          */
-  /************************************************************/
-
-  class GLUI_API GLUI_EditText : public GLUI_Control
-  {
-  public:
-    int                 has_limits;
-    int                 data_type;
-    GLUI_String         orig_text;
-    int                 insertion_pt;
-    int                 title_x_offset;
-    int                 text_x_offset;
-    int                 substring_start; /*substring that gets displayed in box*/
-    int                 substring_end;  
-    int                 sel_start, sel_end;  /* current selection */
-    int                 num_periods;
-    int                 last_insertion_pt;
-    float               float_low, float_high;
-    int                 int_low, int_high;
-    GLUI_Spinner       *spinner;
-    int                 draw_text_only;
-
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  key_handler( unsigned char key,int modifiers );
-    int  special_handler( int key, int modifiers );
-
-    void activate( int how );
-    void deactivate( void );
-
-    void draw( int x, int y );
-
-    int  mouse_over( int state, int x, int y );
-
-    int  find_word_break( int start, int direction );
-    int  substring_width( int start, int end );
-    void clear_substring( int start, int end );
-    int  find_insertion_pt( int x, int y );
-    int  update_substring_bounds( void );
-    void update_and_draw_text( void );
-    void draw_text( int x, int y );
-    void draw_insertion_pt( void );
-    void set_numeric_text( void );
-    void update_x_offsets( void );
-    void update_size( void );
-
-    void set_float_limits( float low,float high,int limit_type=GLUI_LIMIT_CLAMP);
-    void set_int_limits( int low, int high, int limit_type=GLUI_LIMIT_CLAMP );
-    void set_float_val( float new_val );
-    void set_int_val( int new_val );
-    void set_text( const char *text );
-    void set_text( const GLUI_String &s) { set_text(s.c_str()); }
-    const char *get_text()               { return text.c_str(); }
-
-    void dump( FILE *out, const char *text );
-
-    // Constructor, no live variable
-    GLUI_EditText( GLUI_Node *parent, const char *name,
-                   int text_type=GLUI_EDITTEXT_TEXT,
-                   int id=-1, CallBack callback=CallBack() );
-    // Constructor, int live variable
-    GLUI_EditText( GLUI_Node *parent, const char *name,
-                   int *live_var,
-                   int id=-1, CallBack callback=CallBack() );
-    // Constructor, float live variable
-    GLUI_EditText( GLUI_Node *parent, const char *name,
-                   float *live_var,
-                   int id=-1, CallBack callback=CallBack() );
-    // Constructor, char* live variable
-    GLUI_EditText( GLUI_Node *parent, const char *name, 
-                   char *live_var,
-                   int id=-1, CallBack callback=CallBack() );
-    // Constructor, std::string live variable
-    GLUI_EditText( GLUI_Node *parent, const char *name, 
-                   std::string &live_var,
-                   int id=-1, CallBack callback=CallBack() );
-
-    // Deprecated constructor, only called internally
-    GLUI_EditText( GLUI_Node *parent, const char *name,
-                   int text_type, void *live_var,
-                   int id, CallBack callback );
-    // Deprecated constructor, only called internally
-    GLUI_EditText( void ) { common_init(); }
-
-  protected:
-    void common_init( void ) {
-      h                     = GLUI_EDITTEXT_HEIGHT;
-      w                     = GLUI_EDITTEXT_WIDTH;
-      title_x_offset        = 0;
-      text_x_offset         = 55;
-      insertion_pt          = -1;
-      last_insertion_pt     = -1;
-      name                  = "";
-      substring_start       = 0;
-      data_type             = GLUI_EDITTEXT_TEXT;
-      substring_end         = 2;
-      num_periods           = 0;
-      has_limits            = GLUI_LIMIT_NONE;
-      sel_start             = 0;
-      sel_end               = 0;
-      active_type           = GLUI_CONTROL_ACTIVE_PERMANENT;
-      can_activate          = true;
-      spacebar_mouse_click  = false;
-      spinner               = NULL;
-      draw_text_only        = false;
-    }
-    void common_construct( GLUI_Node *parent, const char *name, 
-                           int data_type, int live_type, void *live_var,
-                           int id, CallBack callback );
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               CommandLine class                          */
-  /*                                                          */
-  /************************************************************/
 
 #ifdef _MSC_VER
   // Explicit template instantiation needed for dll
@@ -1663,815 +1418,26 @@ namespace glui {
   template class GLUI_API std::vector<GLUI_String, std::allocator<GLUI_String> >;
 #endif
 
-  class GLUI_API GLUI_CommandLine : public GLUI_EditText
-  {
-  public:
-    typedef GLUI_EditText Super;
-
-    enum { HIST_SIZE = 100 };
-
-    std::vector<GLUI_String> hist_list;
-    int  curr_hist;
-    int  oldest_hist;
-    int  newest_hist;
-    bool commit_flag;
-
-  public:
-    int  key_handler( unsigned char key,int modifiers );
-    int  special_handler(	int key,int modifiers );
-    void deactivate( void );
-
-    virtual const char *get_history( int command_number ) const
-    { return hist_list[command_number - oldest_hist].c_str(); }
-    virtual GLUI_String& get_history_str( int command_number )
-    { return hist_list[command_number - oldest_hist]; }
-    virtual const GLUI_String& get_history_str( int command_number ) const
-    { return hist_list[command_number - oldest_hist]; }
-    virtual void recall_history( int history_number );
-    virtual void scroll_history( int direction );
-    virtual void add_to_history( const char *text );
-    virtual void reset_history( void );
-
-    void dump( FILE *out, const char *text );
-
-
-    GLUI_CommandLine( GLUI_Node *parent, const char *name, void *live_var=NULL,
-                      int id=-1, CallBack callback=CallBack() );
-    GLUI_CommandLine( void ) { common_init(); }
-  protected:
-    void common_init() {
-      hist_list.resize(HIST_SIZE);
-      curr_hist = 0;
-      oldest_hist = 0;
-      newest_hist = 0;
-      commit_flag = false;
-    }
-
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*              RadioGroup class (container)                */
-  /*                                                          */
-  /************************************************************/
-
-  class GLUI_API GLUI_RadioGroup : public GLUI_Control
-  {
-  public:
-    int  num_buttons;
-
-    void draw( int x, int y );
-    void set_name( const char *text );
-    void set_int_val( int int_val ); 
-    void set_selected( int int_val );
-
-    void draw_group( int translate );
-
-    GLUI_RadioGroup( GLUI_Node *parent, int *live_var=NULL,
-                     int user_id=-1,CallBack callback=CallBack() );
-    GLUI_RadioGroup( void ) { common_init(); }
-
-  protected:
-    void common_init( void ) {
-      x_off         = 0;
-      y_off_top     = 0;
-      y_off_bot     = 0;
-      is_container  = true;
-      w             = 300;
-      h             = 300;
-      num_buttons   = 0;
-      name          = "";
-      can_activate  = false;
-      live_type     = GLUI_LIVE_INT;
-    }
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               RadioButton class (container)              */
-  /*                                                          */
-  /************************************************************/
-
-  class GLUI_API GLUI_RadioButton : public GLUI_Control
-  {
-  public:
-    int orig_value;
-    bool currently_inside;
-    int text_x_offset;
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-
-    void draw( int x, int y );
-    void update_size( void );
-
-    void draw_active_area( void );
-    void draw_checked( void );
-    void draw_unchecked( void );
-    void draw_O( void );
-
-    GLUI_RadioButton( GLUI_RadioGroup *group, const char *name );
-    GLUI_RadioGroup *group;
-
-  protected:
-    void common_init()
-    {
-      glui_format_str( name, "RadioButton: %p", (void *) this );
-      h              = GLUI_RADIOBUTTON_SIZE;
-      group          = NULL;
-      orig_value     = -1;
-      text_x_offset  = 18;
-      can_activate   = true;
-    }
-  };
-
-
-  /************************************************************/
-  /*                                                          */
-  /*               Separator class (container)                */
-  /*                                                          */
-  /************************************************************/
-
-  class GLUI_API GLUI_Separator : public GLUI_Control
-  {
-  public:
-    void draw( int x, int y );
-
-    GLUI_Separator( GLUI_Node *parent );
-    GLUI_Separator( void ) { common_init(); }
-
-  protected:
-    void common_init() {
-      w            = 100;
-      h            = GLUI_SEPARATOR_HEIGHT;
-      can_activate = false;
-    }
-  };
-
-#define  GLUI_SPINNER_ARROW_WIDTH   12
-#define  GLUI_SPINNER_ARROW_HEIGHT   8
-#define  GLUI_SPINNER_ARROW_Y        2
-
-#define  GLUI_SPINNER_STATE_NONE     0
-#define  GLUI_SPINNER_STATE_UP       1
-#define  GLUI_SPINNER_STATE_DOWN     2
-#define  GLUI_SPINNER_STATE_BOTH     3
-
-#define  GLUI_SPINNER_DEFAULT_GROWTH_EXP   1.05f
-
-  /************************************************************/
-  /*                                                          */
-  /*               Spinner class (container)                  */
-  /*                                                          */
-  /************************************************************/
  
-  class GLUI_API GLUI_Spinner : public GLUI_Control
-  {
-  public:
-    // Constructor, no live var
-    GLUI_Spinner( GLUI_Node* parent, const char *name, 
-                  int data_type=GLUI_SPINNER_INT, int id=-1, CallBack callback=CallBack() );
-    // Constructor, int live var
-    GLUI_Spinner( GLUI_Node* parent, const char *name, 
-                  int *live_var, int id=-1, CallBack callback=CallBack() );
-    // Constructor, float live var
-    GLUI_Spinner( GLUI_Node* parent, const char *name, 
-                  float *live_var, int id=-1, CallBack callback=CallBack() );
-    // Deprecated constructor
-    GLUI_Spinner( GLUI_Node* parent, const char *name, 
-                  int data_type,
-                  void *live_var,
-                  int id=-1, CallBack callback=CallBack() );
-    // Deprecated constructor
-    GLUI_Spinner( void ) { common_init(); }
-
-    bool          currently_inside;
-    int           state;
-    float         growth, growth_exp;
-    int           last_x, last_y;
-    int           data_type;
-    int           callback_count;
-    int           last_int_val;
-    float         last_float_val;
-    int           first_callback;
-    float         user_speed;
-
-    GLUI_EditText *edittext;
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  key_handler( unsigned char key,int modifiers );
-    int  special_handler(   int key,int modifiers );
-
-    void draw( int x, int y );
-    void draw_pressed( void );
-    void draw_unpressed( void );
-    void draw_text( int sunken );
-
-    void update_size( void );
-
-    void set_float_limits( float low,float high,int limit_type=GLUI_LIMIT_CLAMP);
-    void set_int_limits( int low, int high,int limit_type=GLUI_LIMIT_CLAMP);
-    int  find_arrow( int local_x, int local_y );
-    void do_drag( int x, int y );
-    void do_callbacks( void );
-    void do_click( void );
-    void idle( void );
-    bool needs_idle( void ) const;
-
-    const char *get_text( void );
-
-    void set_float_val( float new_val );
-    void set_int_val( int new_val );
-    float  get_float_val( void );
-    int    get_int_val( void );
-    void increase_growth( void );
-    void reset_growth( void );
-
-    void set_speed( float speed ) { user_speed = speed; }
-
-  protected:
-    void common_init() {
-      glui_format_str( name, "Spinner: %p", this );
-      h            = GLUI_EDITTEXT_HEIGHT;
-      w            = GLUI_EDITTEXT_WIDTH;
-      x_off        = 0;
-      y_off_top    = 0;
-      y_off_bot    = 0;
-      can_activate = true;
-      state        = GLUI_SPINNER_STATE_NONE;
-      edittext     = NULL;
-      growth_exp   = GLUI_SPINNER_DEFAULT_GROWTH_EXP;
-      callback_count = 0;
-      first_callback = true;
-      user_speed   = 1.0;
-    }
-    void common_construct( GLUI_Node* parent, const char *name, 
-                           int data_type, void *live_var,
-                           int id, CallBack callback );
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               StaticText class                           */
-  /*                                                          */
-  /************************************************************/
-
-  class GLUI_API GLUI_StaticText : public GLUI_Control
-  {
-  public:
-    void set_text( const char *text );
-    void draw( int x, int y );
-    void draw_text( void );
-    void update_size( void );
-    void erase_text( void );
-
-    GLUI_StaticText(GLUI_Node *parent, const char *name);
-    GLUI_StaticText( void ) { common_init(); }
-
-  protected:
-    void common_init() {
-      h       = GLUI_STATICTEXT_SIZE;
-      name    = "";
-      can_activate  = false;
-    }
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               TextBox class - JVK                        */
-  /*                                                          */
-  /************************************************************/
-
-  class GLUI_API GLUI_TextBox : public GLUI_Control
-  {
-  public:
-    /* GLUI_Context *Textbox - JVK */
-    GLUI_TextBox(GLUI_Node *parent, GLUI_String &live_var,
-                 bool scroll = false, int id=-1, CallBack callback=CallBack() );
-    GLUI_TextBox( GLUI_Node *parent,
-                  bool scroll = false, int id=-1,
-                  CallBack callback=CallBack() );
-
-    GLUI_String         orig_text;
-    int                 insertion_pt;
-    int                 substring_start; /*substring that gets displayed in box*/
-    int                 substring_end;  
-    int                 sel_start, sel_end;  /* current selection */
-    int                 last_insertion_pt;
-    int                 draw_text_only;
-    int                 tab_width;
-    int                 start_line;
-    int                 num_lines;
-    int                 curr_line;
-    int                 visible_lines;
-    int                 insert_x;        /* Similar to "insertion_pt", these variables keep */
-    int                 insert_y;        /* track of where the ptr is, but in pixels */
-    int                 keygoal_x;       /* where up down keys would like to put insertion pt*/
-    GLUI_Scrollbar     *scrollbar;
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  key_handler( unsigned char key,int modifiers );
-    int  special_handler( int key,int modifiers );
-  
-    void activate( int how );
-    void deactivate( void );
-
-    void enable( void );
-    void disable( void );
-
-    void draw( int x, int y );
-
-    int  mouse_over( int state, int x, int y );
-
-    int get_box_width();
-    int  find_word_break( int start, int direction );
-    int  substring_width( int start, int end, int initial_width=0 );
-    void clear_substring( int start, int end );
-    int  find_insertion_pt( int x, int y );
-    int  update_substring_bounds( void );
-    void update_and_draw_text( void );
-    void draw_text( int x, int y );
-    void draw_insertion_pt( void );
-    void update_x_offsets( void );
-    void update_size( void );
-
-    void set_text( const char *text );
-    const char *get_text( void )         { return text.c_str(); }
-
-    void dump( FILE *out, const char *text );
-    void set_tab_w(int w) { tab_width = w; }
-    void set_start_line(int l) { start_line = l; }
-    static void scrollbar_callback(GLUI_Control*);
-
-    bool wants_tabs( void ) const { return true; }
-
-  protected:
-    void common_init()
-    {
-      h                     = GLUI_TEXTBOX_HEIGHT;
-      w                     = GLUI_TEXTBOX_WIDTH;
-      tab_width             = GLUI_TAB_WIDTH;
-      num_lines             = 0;
-      visible_lines         = 0;
-      start_line            = 0;
-      curr_line             = 0;
-      insert_y              = -1;
-      insert_x              = -1;
-      insertion_pt          = -1;
-      last_insertion_pt     = -1;
-      name[0]               = '\0';
-      substring_start       = 0;
-      substring_end         = 2;
-      sel_start             = 0;
-      sel_end               = 0;
-      active_type           = GLUI_CONTROL_ACTIVE_PERMANENT;
-      can_activate          = true;
-      spacebar_mouse_click  = false;
-      scrollbar             = NULL;
-      draw_text_only        = false;
-    }
-    void common_construct(
-                          GLUI_Node *parent, GLUI_String *live_var, 
-                          bool scroll, int id, CallBack callback); 
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*                   List class - JVK                       */
-  /*                                                          */
-  /************************************************************/
-
-  class GLUI_API GLUI_List_Item : public GLUI_Node 
-  {
-  public:
-    GLUI_String text;
-    int         id;
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               List class - JVK                           */
-  /*                                                          */
-  /************************************************************/
-
-  class GLUI_API GLUI_List : public GLUI_Control
-  {
-  public:
-    /* GLUI_Context *List - JVK */
-    GLUI_List( GLUI_Node *parent, bool scroll = false,
-               int id=-1, CallBack callback=CallBack() );
-    /*, GLUI_Control *object = NULL 
-      ,GLUI_InterObject_CB obj_cb = NULL);*/
-
-    GLUI_List( GLUI_Node *parent,
-               GLUI_String& live_var, bool scroll = false, 
-               int id=-1, 
-               CallBack callback=CallBack()
-               /*,GLUI_Control *object = NULL */
-               /*,GLUI_InterObject_CB obj_cb = NULL*/);
-
-
-    GLUI_String         orig_text;
-    int                 draw_text_only;
-    int                 start_line;
-    int                 num_lines;
-    int                 curr_line;
-    int                 visible_lines;
-    GLUI_Scrollbar      *scrollbar;
-    GLUI_List_Item      items_list;
-    GLUI_Control        *associated_object;
-    CallBack             obj_cb;
-    int                 cb_click_type;
-    int                 last_line;
-    int                 last_click_time;
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  key_handler( unsigned char key,int modifiers );
-    int  special_handler( int key,int modifiers );
-  
-    void activate( int how );
-    void deactivate( void );
-
-    void draw( int x, int y );
-
-    int  mouse_over( int state, int x, int y );
-
-    int get_box_width();
-    int  find_word_break( int start, int direction );
-    int  substring_width( const char *t, int start, int end );
-    int  find_line( int x, int y );
-    void update_and_draw_text( void );
-    void draw_text( const char *t, int selected, int x, int y );
-    void update_size( void );
-
-
-    int  add_item( int id, const char *text );
-    int  delete_item( const char *text );
-    int  delete_item( int id );
-    int  delete_all();
-
-    GLUI_List_Item *get_item_ptr( const char *text );
-    GLUI_List_Item *get_item_ptr( int id );
-
-    void dump( FILE *out, const char *text );
-    void set_start_line(int l) { start_line = l; }
-    static void scrollbar_callback(GLUI_Control*);
-    int get_current_item() { return curr_line; }
-    void set_click_type(int d) {
-      cb_click_type = d; }
-    void set_object_callback(CallBack cb=CallBack(), GLUI_Control*obj=NULL)
-    { obj_cb=cb; associated_object=obj; }
-
-  protected:
-    void common_init()
-    {
-      h                     = GLUI_LIST_HEIGHT;
-      w                     = GLUI_LIST_WIDTH;
-      num_lines             = 0;
-      visible_lines         = 0;
-      start_line            = 0;
-      curr_line             = 0;
-      name[0]               = '\0';
-      active_type           = GLUI_CONTROL_ACTIVE_PERMANENT;
-      can_activate          = true;
-      spacebar_mouse_click  = false;
-      scrollbar             = NULL;
-      draw_text_only        = false;
-      cb_click_type         = GLUI_SINGLE_CLICK;
-      last_line             = -1;
-      last_click_time       = 0;
-      associated_object     = NULL;
-    };
-    void common_construct(
-                          GLUI_Node *parent,
-                          GLUI_String* live_var, bool scroll,
-                          int id,
-                          CallBack callback
-                          /*,GLUI_Control *object*/
-                          /*,GLUI_InterObject_CB obj_cb*/);
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*               Scrollbar class - JVK                      */
-  /*                                                          */
-  /************************************************************/
- 
-  class GLUI_API GLUI_Scrollbar : public GLUI_Control
-  {
-  public:
-    // Constructor, no live var
-    GLUI_Scrollbar( GLUI_Node *parent,
-                    const char *name, 
-                    int horz_vert=GLUI_SCROLL_HORIZONTAL,
-                    int data_type=GLUI_SCROLL_INT,
-                    int id=-1, CallBack callback=CallBack() 
-                    /*,GLUI_Control *object = NULL*/
-                    /*,GLUI_InterObject_CB obj_cb = NULL*/
-                    );
-
-    // Constructor, int live var
-    GLUI_Scrollbar( GLUI_Node *parent, const char *name, int horz_vert,
-                    int *live_var,
-                    int id=-1, CallBack callback=CallBack() 
-                    /*,GLUI_Control *object = NULL*/
-                    /*,GLUI_InterObject_CB obj_cb = NULL*/
-                    );
-
-    // Constructor, float live var
-    GLUI_Scrollbar( GLUI_Node *parent, const char *name, int horz_vert,
-                    float *live_var,
-                    int id=-1, CallBack callback=CallBack()
-                    /*,GLUI_Control *object = NULL*/
-                    /*,GLUI_InterObject_CB obj_cb = NULL*/
-                    );
-
-    bool          currently_inside;
-    int           state;
-    float         growth, growth_exp;
-    int           last_x, last_y;
-    int           data_type;
-    int           callback_count;
-    int           last_int_val;  ///< Used to prevent repeated callbacks.
-    float         last_float_val;
-    int           first_callback;
-    float         user_speed;
-    float         float_min, float_max;
-    int           int_min, int_max;
-    int           horizontal;
-    double     last_update_time; ///< GLUI_Time() we last advanced scrollbar.
-    double     velocity_limit; ///< Maximum distance to advance per second.
-    int box_length;
-    int box_start_position;
-    int box_end_position;
-    int track_length;
-
-
-    /* Rather than directly access an Editbox or Textbox for 
-       changing variables, a pointer to some object is defined
-       along with a static callback in the form func(void *, int) -
-       the int is the new value, the void * must be cast to that
-       particular object type before use.
-    */
-    void *        associated_object; /* Lets the Spinner manage it's own callbacks */
-    CallBack       object_cb; /* function pointer to object call_back */
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  key_handler( unsigned char key,int modifiers );
-    int  special_handler( int key,int modifiers );
-  
-    void draw( int x, int y );
-    void draw_pressed( void );
-    void draw_unpressed( void );
-    void draw_text( int sunken );
-
-    void update_size( void );
-
-    void set_int_limits( int low, int high,int limit_type=GLUI_LIMIT_CLAMP);
-    void set_float_limits( float low,float high,int limit_type=GLUI_LIMIT_CLAMP);
-    int  find_arrow( int local_x, int local_y );
-    void do_drag( int x, int y );
-    void do_callbacks( void );
-    void draw_scroll( void );
-    void do_click( void );
-    void idle( void );
-    bool needs_idle( void ) const;
-    void set_int_val( int new_val );
-    void set_float_val( float new_val );
-    void increase_growth( void );
-    void reset_growth( void );
-
-    void set_speed( float speed ) { user_speed = speed; };
-    void update_scroll_parameters();
-    void set_object_callback(CallBack cb=CallBack(), GLUI_Control*obj=NULL)
-    { object_cb=cb; associated_object=obj; }
-
-  protected:
-    void common_init ( void );
-    void common_construct(
-                          GLUI_Node *parent,
-                          const char *name, 
-                          int horz_vert,
-                          int data_type, void* live_var,
-                          int id, CallBack callback
-                          /*,GLUI_Control *object
-                            ,GLUI_InterObject_CB obj_cb*/
-                          );
-
-    virtual void draw_scroll_arrow(int arrowtype, int x, int y);
-    virtual void draw_scroll_box(int x, int y, int w, int h);
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*                   Listbox class                          */
-  /*                                                          */
-  /************************************************************/
-
-  class GLUI_API GLUI_Listbox_Item : public GLUI_Node 
-  {
-  public:
-    GLUI_String text;
-    int         id;
-  };
-
-  class GLUI_API GLUI_Listbox : public GLUI_Control
-  {
-  public:
-    GLUI_String       curr_text;
-    GLUI_Listbox_Item items_list;
-    int               depressed;
-
-    int  orig_value;
-    bool currently_inside;
-    int  text_x_offset, title_x_offset;
-    int  glut_menu_id;
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  key_handler( unsigned char key,int modifiers );
-    int  special_handler( int key,int modifiers );
-
-    void update_size( void );
-    void draw( int x, int y );
-    int  mouse_over( int state, int x, int y );
-
-    void set_int_val( int new_val );
-    void dump( FILE *output );
-
-    int  add_item( int id, const char *text );
-    int  delete_item( const char *text );
-    int  delete_item( int id );
-    int  sort_items( void );
-
-    int  do_selection( int item );
-
-    GLUI_Listbox_Item *get_item_ptr( const char *text );
-    GLUI_Listbox_Item *get_item_ptr( int id );
-  
-
-    GLUI_Listbox( GLUI_Node *parent,
-                  const char *name, int *live_var=NULL,
-                  int id=-1, CallBack callback=CallBack() );
-    GLUI_Listbox( void ) { common_init(); }
-
-  protected:
-    /** Change w and return true if we need to be widened to fit the current item. */
-    bool recalculate_item_width( void );
-    void common_init() {
-      glui_format_str( name, "Listbox: %p", this );
-      w              = GLUI_EDITTEXT_WIDTH;
-      h              = GLUI_EDITTEXT_HEIGHT;
-      orig_value     = -1;
-      title_x_offset = 0;
-      text_x_offset  = 55;
-      can_activate   = true;
-      curr_text      = "";
-      live_type      = GLUI_LIVE_INT;  /* This has an integer live var */
-      depressed      = false;
-      glut_menu_id   = -1;
-    }
-
-    ~GLUI_Listbox();
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*              Mouse_Interaction class                     */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     This is the superclass of translation and rotation widgets.
-  */
-  class GLUI_API GLUI_Mouse_Interaction : public GLUI_Control
-  {
-  public:
-    /*int  get_main_area_size( void ) { return MIN( h-18,  */
-    int            draw_active_area_only;
-
-    int  mouse_down_handler( int local_x, int local_y );
-    int  mouse_up_handler( int local_x, int local_y, bool inside );
-    int  mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  special_handler( int key, int modifiers );
-    void update_size( void );
-    void draw( int x, int y );
-    void draw_active_area( void );
-
-    /***  The following methods (starting with "iaction_") need to
-          be overloaded  ***/
-    virtual int  iaction_mouse_down_handler( int local_x, int local_y ) = 0;
-    virtual int  iaction_mouse_up_handler( int local_x, int local_y, bool inside )=0;
-    virtual int  iaction_mouse_held_down_handler( int local_x, int local_y, bool inside )=0;
-    virtual int  iaction_special_handler( int key, int modifiers )=0;
-    virtual void iaction_draw_active_area_persp( void )=0;
-    virtual void iaction_draw_active_area_ortho( void )=0;
-    virtual void iaction_dump( FILE *output )=0;
-    virtual void iaction_init( void ) = 0;
-  
-    GLUI_Mouse_Interaction( void ) {
-      glui_format_str( name, "Mouse_Interaction: %p", this );
-      w              = GLUI_MOUSE_INTERACTION_WIDTH;
-      h              = GLUI_MOUSE_INTERACTION_HEIGHT;
-      can_activate   = true;
-      live_type      = GLUI_LIVE_NONE;
-      alignment      = GLUI_ALIGN_CENTER;
-      draw_active_area_only = false;
-    }
-  };
-
-  /************************************************************/
-  /*                                                          */
-  /*                   Rotation class                         */
-  /*                                                          */
-  /************************************************************/
-
-  /**
-     An onscreen rotation controller--allows the user to interact with
-     a 3D rotation via a spaceball-like interface.
-  */
-  class GLUI_API GLUI_Rotation : public GLUI_Mouse_Interaction
-  {
-  public:
-    Arcball        *ball;
-    GLUquadricObj *quadObj;
-    bool           can_spin, spinning;
-    float          damping;
-  
-    int  iaction_mouse_down_handler( int local_x, int local_y );
-    int  iaction_mouse_up_handler( int local_x, int local_y, bool inside );
-    int  iaction_mouse_held_down_handler( int local_x, int local_y, bool inside );
-    int  iaction_special_handler( int key, int modifiers );
-    void iaction_init( void ) { init_ball(); }
-    void iaction_draw_active_area_persp( void );
-    void iaction_draw_active_area_ortho( void );
-    void iaction_dump( FILE *output );
-
-    /*  void update_size( void ); */
-    /*  void draw( int x, int y ); */
-    /*  int mouse_over( int state, int x, int y ); */
-
-    void setup_texture( void );
-    void setup_lights( void );
-    void draw_ball( float radius );
-
-    void init_ball( void );
-
-    void reset( void );
-
-    bool needs_idle( void ) const;
-    void idle( void );
-
-    void copy_float_array_to_ball( void );
-    void copy_ball_to_float_array( void );
-
-    void set_spin( float damp_factor );
-
-    GLUI_Rotation( GLUI_Node *parent, const char *name, float *live_var=NULL,
-                   int id=-1, CallBack callback=CallBack() );
-    GLUI_Rotation(void) { common_init(); }
-
-  protected:
-    void common_init();
-  };
 
   /********** Misc functions *********************/
   int _glutBitmapWidthString( void *font, const char *s );
   void _glutBitmapString( void *font, const char *s );
-
-  /********** Our own callbacks for glut *********/
-  /* These are the callbacks that we pass to glut.  They take
-     some action if necessary, then (possibly) call the user-level
-     glut callbacks.  
-  */
-
-  // void glui_display_func( void );
-  // void glui_reshape_func( int w, int h );
-  // void glui_keyboard_func(unsigned char key, int x, int y);
-  // void glui_special_func(int key, int x, int y);
-  // void glui_mouse_func(int button, int state, int x, int y);
-  // void glui_motion_func(int x, int y);
-  // void glui_passive_motion_func(int x, int y);
-  // void glui_entry_func(int state);
-  // void glui_visibility_func(int state);
-  // void glui_idle_func(void);
-
-  // void glui_parent_window_reshape_func( int w, int h );
-  // void glui_parent_window_keyboard_func(unsigned char key, int x, int y);
-  // void glui_parent_window_mouse_func(int, int, int, int );
-  // void glui_parent_window_special_func(int key, int x, int y);
-
 }
 
 
+#include "glui/Context.h"
+#include "glui/Spinner.h"
+#include "glui/Separator.h"
+#include "glui/RadioButton.h"
+#include "glui/RadioGroup.h"
+#include "glui/EditText.h"
+#include "glui/CommandLine.h"
+#include "glui/StaticText.h"
+#include "glui/TextBox.h"
+#include "glui/List.h"
+#include "glui/ScrollBar.h"
+#include "glui/ListBox.h"
+#include "glui/MouseInteraction.h"
 #include "glui/Translation.h"
+#include "glui/Rotation.h"
