+++ glui/include/GL/glui/Separator.h
@@ -1,0 +1,54 @@
+/****************************************************************************
+
+  GLUI_Context *User Interface Toolkit
+  ---------------------------
+
+     glui.h - Main (and only) external header for 
+        GLUI_Context *User Interface Toolkit
+
+          --------------------------------------------------
+
+  Copyright (c) 1998 Paul Rademacher
+
+  WWW:    http://sourceforge.net/projects/glui/
+  Forums: http://sourceforge.net/forum/?group_id=92496
+
+  This software is provided 'as-is', without any express or implied 
+  warranty. In no event will the authors be held liable for any damages 
+  arising from the use of this software. 
+
+  Permission is granted to anyone to use this software for any purpose, 
+  including commercial applications, and to alter it and redistribute it 
+  freely, subject to the following restrictions: 
+
+  1. The origin of this software must not be misrepresented; you must not 
+  claim that you wrote the original software. If you use this software 
+  in a product, an acknowledgment in the product documentation would be 
+  appreciated but is not required. 
+  2. Altered source versions must be plainly marked as such, and must not be 
+  misrepresented as being the original software. 
+  3. This notice may not be removed or altered from any source distribution. 
+
+*****************************************************************************/
+
+#pragma once
+
+namespace glui {
+
+  class GLUI_API GLUI_Separator : public GLUI_Control
+  {
+  public:
+    void draw( int x, int y );
+
+    GLUI_Separator( GLUI_Node *parent );
+    GLUI_Separator( void ) { common_init(); }
+
+  protected:
+    void common_init() {
+      w            = 100;
+      h            = GLUI_SEPARATOR_HEIGHT;
+      can_activate = false;
+    }
+  };
+
+}
