+++ glui/include/GL/glui/StaticText.h
@@ -1,0 +1,58 @@
+/****************************************************************************
+
+  GLUI_Context *User Interface Toolkit
+  ---------------------------
+
+     glui.h - Main (and only) external header for 
+        GLUI_Context *User Interface Toolkit
+
+          --------------------------------------------------
+
+  Copyright (c) 1998 Paul Rademacher
+
+  WWW:    http://sourceforge.net/projects/glui/
+  Forums: http://sourceforge.net/forum/?group_id=92496
+
+  This software is provided 'as-is', without any express or implied 
+  warranty. In no event will the authors be held liable for any damages 
+  arising from the use of this software. 
+
+  Permission is granted to anyone to use this software for any purpose, 
+  including commercial applications, and to alter it and redistribute it 
+  freely, subject to the following restrictions: 
+
+  1. The origin of this software must not be misrepresented; you must not 
+  claim that you wrote the original software. If you use this software 
+  in a product, an acknowledgment in the product documentation would be 
+  appreciated but is not required. 
+  2. Altered source versions must be plainly marked as such, and must not be 
+  misrepresented as being the original software. 
+  3. This notice may not be removed or altered from any source distribution. 
+
+*****************************************************************************/
+
+#pragma once
+
+namespace glui {
+
+  class GLUI_API GLUI_StaticText : public GLUI_Control
+  {
+  public:
+    void set_text( const char *text );
+    void draw( int x, int y );
+    void draw_text( void );
+    void update_size( void );
+    void erase_text( void );
+
+    GLUI_StaticText(GLUI_Node *parent, const char *name);
+    GLUI_StaticText( void ) { common_init(); }
+
+  protected:
+    void common_init() {
+      h       = GLUI_STATICTEXT_SIZE;
+      name    = "";
+      can_activate  = false;
+    }
+  };
+  
+}
